[Data] Move legacy data usage collection to usage/util.py
